requirements.txt
streamlit~=1.4.0
pandas~=1.3.2
Pillow~=7.2.0
streamlit_option_menu